Merge branch 'master' into 'master'
Addresses regression with jenkins setup that does not use the multiproject setup

The Jenkins Gitlab Hook Plugin allows for two kinds of setups.  One where a Jenkins job is created for each branch and one where all branches are built in a single Jenkins job.  MR !9 added support for the multi-job setup but broke the single job setup.  This request addresses #30 and fixes the regression.

See merge request !27