Merge branch 'fix_ldap_group_sync' into 'master'
Downgrade group member role if LDAP dictates

Fixes #170 

If a group link was updated to set a lower maximum role/access level, 
a user's role was not downgraded. Similarly, if the user was moved to
another group where their role should be lowered, their access was not
downgraded. For some reason we were preferring the higher of LDAP or
GitLab access even if LDAP said it should be lower. After this change,
what LDAP says, wins. 

If a user is a member of multiple LDAP groups that are linked with the
same GitLab group, they still receive the highest of all roles.

See merge request !123