Merge branch 'increase_lfs_size_column' into 'master'
Increase LFS objects size column

Fixes #12745

Increases the `size` column integer limit to an 8-byte integer. This allows for a max value of `9223372036854775807` which is 9,223,372,036 GB. That should do it :smiley:

I tested this by first reproducing the error (push a file larger than 2.1 GB). The error was:

```
RangeError (3145728000 is out of range for ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Integer with limit 4):
  lib/gitlab/lfs/response.rb:232:in `store_file'
  lib/gitlab/lfs/response.rb:170:in `render_lfs_upload_ok'
  lib/gitlab/lfs/response.rb:51:in `block in render_storage_upload_store_response'
  lib/gitlab/lfs/response.rb:204:in `render_response_to_push'
  lib/gitlab/lfs/response.rb:50:in `render_storage_upload_store_response'
  lib/gitlab/lfs/router.rb:76:in `put_response'
  lib/gitlab/lfs/router.rb:20:in `try_call'
  lib/gitlab/backend/grack_auth.rb:41:in `call'
  lib/gitlab/backend/grack_auth.rb:18:in `call_with_kerberos_support'
  lib/gitlab/backend/grack_auth.rb:8:in `call'
```

Then I ran this migration and pushed the file again. It uploaded successfully.

See merge request !2644