Merge branch 'gh-import-ssl-verification' into 'master'
Allow OAuth SSL verification to be disabled when importing from GitHub

Close https://gitlab.com/gitlab-org/gitlab-ee/issues/457

The SSL verification should be disabled if the GitHub Enterprise use a self signed certificate. Even disabling this on GitLab, the Git client installed in the server will perform the verification and stop the process if it is unknown. So, we need to disable Git SSL verification in the server hosting `GitLab` with the following command:

```
$ git config --global http.sslVerify false
```

/cc @balameb @ernstvn

See merge request !323

Signed-off-by: Rémy Coutable <remy@rymai.me>