Merge branch 'cherry-pick-401d3a48' into 'master'
pages: Fix "undefined local variable or method `total_size'" when maximum page size is exceeded

Seems like a simple fix - total_size is unqualified in the error message, leading to a NoMethodError being raised instead of the expected error. 

Suggests some tests are missing, but I'm on work time and in a bit of a hurry!

I suspect there's an underlying issue: The admin site says that a maximum pages size of 0 is "unlimited", so I set it to 0. This error then arose trying to build pages, implying 0 actually means 0 in this context at the moment.


See merge request !356