Merge branch 'ldap_fix_api' into 'master'
Fix bad Group class resolution when querying LDAP via API

When querying LDAP groups via the API, there was a 500 error with a strange stack. Thanks to @DouweM we figured out that in the context of the API, it's choosing `::Group` instead of `EE::Gitlab::LDAP::Group`. This makes it just specific enough to work. 

See merge request !572