Merge branch '748-update-mirror' into 'master'
Fix race condition with UpdateMirrorWorker Lease

Any time we modify the project import state to an in progress we're going to execute the sidekiq job not lease to try to obtain in this workflow. This fixes the UI problem when an user try to update a mirror just inside a lease timeout.

Closes #748

See merge request !641
(cherry picked from commit 38f7f31a9da55ee3b267ed20390856b294ca4f55)