Merge branch '1128-geo-custom-avatars' into 'master'
Geo: Display Custom Avatars in secondary nodes

This MR introduces some small patches for `User`, `Project` and `Group` models and related classes to build the avatar URL pointing to the primary node.
We should be able to remove this in the future when we start replicating local assets in Geo (#76).

For the changes to take effect on all pages, it's required to run:

```
gitlab-rake cache:clear

RAILS_ENV=production bundle exec rake cache:clear
```

Closes #1128

See merge request !904